refactoring of files and folder structure in src
