+++ perf/tabulate_perf_summary.jl
@@ -2,7 +2,7 @@ import Plots
 import Dates
 import JSON
 import OrderedCollections
-include(joinpath(@__DIR__, "..", "src", "utils", "yaml_helper.jl"))
+include(joinpath(@__DIR__, "..", "src", "solver", "yaml_helper.jl"))
 
 function sorted_dataset_folder(; dir = pwd())
     matching_paths = String[]
