+++ src/prognostic_equations/dss.jl
@@ -1,0 +1,24 @@
+using LinearAlgebra: ×, norm, dot
+
+import ClimaAtmos.Parameters as CAP
+using ClimaCore: Operators, Fields, Limiters, Geometry, Spaces
+
+import ClimaComms
+using ClimaCore.Geometry: ⊗
+
+import Thermodynamics as TD
+
+using ClimaCore.Utilities: half
+
+import ClimaCore.Fields: ColumnField
+
+function dss!(Y, p, t)
+    if p.do_dss
+        Spaces.weighted_dss_start2!(Y.c, p.ghost_buffer.c)
+        Spaces.weighted_dss_start2!(Y.f, p.ghost_buffer.f)
+        Spaces.weighted_dss_internal2!(Y.c, p.ghost_buffer.c)
+        Spaces.weighted_dss_internal2!(Y.f, p.ghost_buffer.f)
+        Spaces.weighted_dss_ghost2!(Y.c, p.ghost_buffer.c)
+        Spaces.weighted_dss_ghost2!(Y.f, p.ghost_buffer.f)
+    end
+end
